Update: 상대픽 구름 추가

+++ libs/m/src/client/components/splash.ts
@@ -1,21 +1,39 @@
 import { Easing, Tween } from '@tweenjs/tween.js';
 import { JobId } from '../../common/enums';
+import { getJob } from '../../common/jobs';
 
-export class Splash extends PIXI.Sprite {
+export class Splash extends PIXI.Container {
   private _jobId: JobId = JobId.BEGINNER;
+  private _scale: number;
+  private _isOpponent: boolean;
+  private _direction: 'left' | 'right';
 
   sprite: PIXI.heaven.Sprite;
+  x: number;
+  y: number;
 
-  constructor(jobId: JobId, x: number, y: number) {
+  constructor(
+    jobId: JobId,
+    x: number,
+    y: number,
+    scale: number,
+    isOpponent = false,
+    direction: 'left' | 'right' = 'left'
+  ) {
     super();
     this.sprite = this.addChild(new PIXI.Sprite()).convertToHeaven();
-    this.sprite.scale.set(2);
+    this.sprite.scale.set(scale);
     this.sprite.anchor.set(0.5);
     this.sprite.position.set(x, y);
     this.jobId = jobId;
+    this.x = x;
+    this.y = y;
+    this._scale = scale;
+    this._isOpponent = isOpponent;
+    this._direction = direction;
   }
 
-  remove = (onRemove: () => void) => {
+  remove = (onRemove?: () => void) => {
     new Tween({
       dark: 0,
     })
@@ -25,14 +43,24 @@ export class Splash extends PIXI.Sprite {
         this.sprite.color.setDark(object.dark, object.dark, object.dark);
       })
       .chain(
-        new Tween({ scale: 2 })
+        new Tween({ scale: this._scale })
           .to({ scale: 0 }, 300)
           .easing(Easing.Quartic.In)
           .onUpdate((object) => {
             this.sprite.scale.set(object.scale);
+            if (this._isOpponent) {
+              const job = getJob(this.jobId);
+              this.sprite.scale.x *=
+                (this._direction == 'left' && job.reverse) ||
+                (this._direction == 'right' && !job.reverse)
+                  ? -1
+                  : 1;
+            }
           })
           .onComplete(() => {
-            onRemove();
+            if (onRemove) {
+              onRemove();
+            }
           })
       )
       .start();
@@ -43,13 +71,33 @@ export class Splash extends PIXI.Sprite {
   }
 
   set jobId(value: JobId) {
-    if (!value || this._jobId == value) {
+    if (this._jobId == value) {
       return;
     }
     this._jobId = value;
 
-    this.sprite.texture = PIXI.Texture.from(
-      `../assets/splashes/${this.jobId}.png`
-    );
+    if (this._jobId == 0) {
+      this.sprite.texture = PIXI.Texture.EMPTY;
+    } else {
+      this.sprite.texture = PIXI.Texture.from(
+        `../assets/splashes/${this.jobId}.png`
+      );
+
+      if (this._isOpponent) {
+        const job = getJob(this.jobId);
+
+        this.sprite.scale.set(this._scale);
+        this.sprite.scale.x *=
+          (this._direction == 'left' && job.reverse) ||
+          (this._direction == 'right' && !job.reverse)
+            ? -1
+            : 1;
+        this.sprite.anchor.set(
+          0.5 + job.offsetX / 1024,
+          0.5 + job.offsetY / 604
+        );
+        this.sprite.position.set(this.x, this.y);
+      }
+    }
   }
 }
